+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Fixed
+* Fixed an issue where `.camdo` and legacy `.wimdo` models did not correctly remap indices for texture assignments.
+
 ## 0.11.0 - 2024-07-19
 ### Added
 * Added rendering support for layered normal maps to xc3_wgpu.